refactor: renamed base menu and rewired dependent components
