+++ components/02-molecules/menu/_menu-item.twig
@@ -27,7 +27,7 @@
   'title': item.title,
   }) %}
 
-{% import "@molecules/menus/_menu.twig" as menus %}
+{% import "@molecules/menu/menu.twig" as menus %}
 {% block menu_item_content %}{% endblock %}
 {% extends "@atoms/lists/_list-item.twig" %}
 {% block list_item_content %}
