+++ components/02-molecules/menu/_menu-list.twig
@@ -9,7 +9,7 @@
 {% extends "@atoms/lists/ul.twig" %}
 {% block list_content %}
   {% for item in items %}
-    {% include "@molecules/menus/_menu-item.twig" with {
+    {% include "@molecules/menu/_menu-item.twig" with {
       li_base_class: item_base_class,
       li_modifiers: item_modifiers,
       li_blockname: item_blockname,
