+++ components/02-molecules/menu/menu.stories.js
@@ -1,0 +1,10 @@
+import menu from './menu.twig';
+
+import menuData from './menu.yml';
+
+/**
+ * Storybook Definition.
+ */
+export default { title: 'Molecules/Menu' };
+
+export const baseMenu = () => menu(menuData);
