+++ components/02-molecules/pager/pager.stories.js
@@ -4,6 +4,7 @@ import pagerData from './pager.yml';
 import pagerEllipsesData from './pager-ellipses.yml';
 import pagerPrevEllipsesData from './pager-prev-ellipses.yml';
 import pagerBothEllipsesData from './pager-both-ellipses.yml';
+import pagerFirstLastData from './pager-first-last.yml';
 
 /**
  * Storybook Definition.
@@ -18,3 +19,6 @@ export const withBoth = () => pager({ ...pagerData, ...pagerBothEllipsesData });
 
 export const withPrevious = () =>
   pager({ ...pagerData, ...pagerPrevEllipsesData });
+
+export const withFirstAndLast = () =>
+  pager({ ...pagerData, ...pagerFirstLastData });
