+++ components/03-organisms/menus/inline/inline-menu.twig
@@ -1,4 +1,4 @@
-{% include "@molecules/menus/_menu.twig" with {
+{% include "@molecules/menu/menu.twig" with {
   menu_class: 'inline-menu',
   items: inline_menu_items,
 } %}
