+++ components/03-organisms/menus/main-menu/main-menu.twig
@@ -15,7 +15,7 @@
     </span>
   </a>
   <div id="main-nav" class="main-nav">
-    {% include "@molecules/menus/_menu.twig" with {
+    {% include "@molecules/menu/menu.twig" with {
       menu_class: "main-menu",
       items: menu_items,
     } %}
