+++ components/03-organisms/menus/social/_social-menu-item.twig
@@ -4,5 +4,5 @@
 
 {% extends "@atoms/lists/_list-item.twig" %}
   {% block list_item_content %}
-    {% include "@molecules/menus/social/_social-menu-link.twig" %}
+    {% include "@organisms/menus/social/_social-menu-link.twig" %}
   {% endblock %}
