+++ components/03-organisms/menus/social/social-menu.twig
@@ -8,6 +8,6 @@
 {% extends "@atoms/lists/ul.twig" %}
   {% block list_content %}
     {% for item in social_menu_items %}
-      {% include "@molecules/menus/social/_social-menu-item.twig" %}
+      {% include "@organisms/menus/social/_social-menu-item.twig" %}
     {% endfor %}
   {% endblock %}
