+++ components/04-templates/layouts.stories.js
@@ -1,9 +1,9 @@
 import fullWidthTwig from './full-width.twig';
 import withSidebarTwig from './with-sidebar.twig';
 
-import mainMenu from '../02-molecules/menus/main-menu/main-menu.yml';
-import socialMenu from '../02-molecules/menus/social/social-menu.yml';
-import footerMenu from '../02-molecules/menus/inline/inline-menu.yml';
+import mainMenu from '../03-organisms/menus/main-menu/main-menu.yml';
+import socialMenu from '../03-organisms/menus/social/social-menu.yml';
+import footerMenu from '../03-organisms/menus/inline/inline-menu.yml';
 
 /**
  * Storybook Definition.
